Move database config to config dir
